refactor(channel): split lib.rs into error, channel, and serve modules
lib.rs carried four concerns in 516 lines while every other part of
this crate already has its own file, plus twelve separate
#[cfg(feature = "client")] attributes that now collapse into one per
mod/pub-use line.

- src/error.rs: ChannelError and its Display/core::error::Error impls.
- src/channel.rs: Channel, its inherent methods, and into_halves.
- src/serve.rs: Shepherd, serve, start, both advice constants, warn,
  and the two thread closures, unchanged.
- lib.rs: crate docs, module declarations, and re-exports only.

Pure move: same items, same bodies, same 28 tests (moved with the
code they test), same public paths via pub use. Three things beyond
literal relocation, all mechanical consequences of the split rather
than behavior changes:

- Two intra-doc links now cross a module boundary and needed
  qualifying (`crate::Channel::open`, `crate::serve`) to keep
  resolving under rustdoc's normal per-module scoping; verified
  against a scratch crate before touching the real one.
- Channel's three fields move from private to pub(crate): they were
  reachable from session.rs's a_channel_over_a_socketpair_round_trips
  test before the split only because Channel and session were both
  directly in lib.rs (private items are visible to the defining
  module and its descendants, and session.rs descends from the crate
  root either way) -- moving Channel out from under session.rs's
  common ancestor would otherwise break that test's direct struct
  literal.
- serve.rs's test module drops the redundant `feature = "client"`
  half of its cfg, since the whole file is already gated at the mod
  declaration in lib.rs.
